[romprefix] Provide indication of successful call to install_prealloc
The existing "gPXE starting execution" message indicates that the BEV
(or INT19) was invoked, but gives no indication on whether or not the
gPXE source was successfully retrieved (e.g. from PMM).  Split the
"starting execution message" into "starting execution...ok"; the "ok"
indicates that the main gPXE body was successfully decompressed and
relocated.

Signed-off-by: Michael Brown <mcb30@ipxe.org>
Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>