[settings] Support IPv6 addresses and allow IPv6 nameservers in DNS
This will allow DHCPv6 to assign DNS nameservers for lookups.
I have chosen to allow DNS to perform lookups to either an IPv6 server
or an IPv4 server to avoid adding too much IPv6-dependent code to DNS.

When assigning an IPv6 address via the "ip6" setting, note that for most
uses a routable prefix (eg, 64) and gateway must be provided as well.

Signed-off-by: Matthew Iselin <matthew@theiselins.net>
Signed-off-by: Marty Connor <mdc@etherboot.org>