[ipv6] Rework option parsing for router solicitation and neighbour advertisement handling
Instead of using an "offset" variable, use iob_pull and iob_len to
retrieve options from the packet buffer.

Also, neighbour advertisements no longer assume anything about the
layout of the packet that has been received. Previously, if any other
option was present before the link-layer address, the assertions would
fail and the packet would be ignored. Now, the option list is
traversed and the correct options are used as necessary.

Signed-off-by: Matthew Iselin <matthew@theiselins.net>
Signed-off-by: Marty Connor <mdc@etherboot.org>