[ipv6] Change return code handling from router solicits
Using 'rc' for flags causes "Error <code>" to be displayed when monojob_wait
completes, which is not very user-friendly. By adding the flags to struct
rsolicit_info we can pass this information around without having to special-case
return code handling.

DHCPv6 has been updated accordingly, and has also been updated to take a
pointer to a struct rsolicit_info, which reduces the number of duplicate
solicitations we do in a normal startup sequence.

Signed-off-by: Matthew Iselin <matthew@theiselins.net>
Signed-off-by: Marty Connor <mdc@etherboot.org>