[ipv6] Verify destination address in ICMP6 and ensure I/O buffer is freed
Previously gPXE would happily respond to any ICMP6 echo packet sent on
the network, regardless of whether that packet was addressed to it or
not. This will stop that from happening.

Also, the previous handling would not free the passed I/O buffer; this
has been fixed in this commit.

Signed-off-by: Matthew Iselin <matthew@theiselins.net>
Signed-off-by: Marty Connor <mdc@etherboot.org>