[ipv6] Check for a valid destination address in ipv6_rx
Previously, gPXE would happily process any IPv6 packet transmitted to
it, even if it was not targeted to the gPXE station. This caused
issues with ICMPv6 echo requests in particular, but also wasted
processing time on other traffic on the network.

Now, the destination address is verified before too much processing is done.

Signed-off-by: Matthew Iselin <matthew@theiselins.net>
Signed-off-by: Marty Connor <mdc@etherboot.org>